⛵️🌲 Updated with Glitch

+++ Data/Commands/say.js
@@ -1,5 +1,9 @@
 let methods = {
   run : async function(client, args, message) {
+    if (message.author.id != process.env.ownerID) {
+      message.reply('This is a owner only command');
+      return;
+    }
     await message.delete();
     message.channel.send(args.join(' '));
     return;
